refactor: Simplify observer initialization and reorganize components
- Moved observer setup logic directly into ObserversInitializer component
- Removed intermediate useObserversSetup hook (less abstraction)
- Renamed ObserversProvider to ObserversInitializer (more accurate)
- Changed ObserversInitializer to be a sibling component instead of wrapper
- Moved components from providers/ to initializers/ directory
- Cleaner architecture with initializers as setup components that return null
