--- frontend/app/components/providers/ObserversProvider.tsx
@@ -1,13 +1,0 @@
-import React, { memo } from "react";
-import { useObserversSetup } from "@/app/hooks/useObserversSetup";
-
-export const ObserversProvider = memo(
-  ({ children }: { children: React.ReactNode }) => {
-    console.log("ObserversProvider rendered");
-    useObserversSetup();
-
-    return <>{children}</>;
-  },
-);
-
-ObserversProvider.displayName = "ObserversProvider";
