+++ frontend/app/stores/useTutorialStore.ts
@@ -101,7 +101,6 @@ export const useTutorialStore = create<TutorialState>((set, get) => ({
   },
 
   advanceStep: () => {
-    /*
     set((state) => {
       const keys = Object.keys(tutorialConfig) as TutorialStep[];
       const next = Math.min(state.stepIndex + 1, keys.length - 1);
@@ -116,7 +115,6 @@ export const useTutorialStore = create<TutorialState>((set, get) => ({
         isTutorialActive: !isCompleted, // Deactivate tutorial when completed
       };
     });
-    */
 
     // Save progress after advancing
     get().saveTutorialProgress();
@@ -143,7 +141,7 @@ export const useTutorialStore = create<TutorialState>((set, get) => ({
   },
 
   setVisible: (visible: boolean) => {
-    // set({ visible });
+    set({ visible });
     get().saveTutorialProgress();
   },
